refactor: Move envoy admin client into admin subpackage with separate config_dump types
Moves agent/internal/envoy/ → agent/internal/envoy/admin/ and splits
the config_dump JSON structs + parser into config_dump.go. Renames
AdminClient → Client since the package name already provides context.
