+++ agent/internal/cni/server/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     importpath = "github.com/bpalermo/aether/agent/internal/cni/server",
     visibility = ["//agent:__subpackages__"],
     deps = [
-        "//agent/internal/envoy",
+        "//agent/internal/envoy/admin",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
         "//agent/internal/xds/proxy",
@@ -41,7 +41,7 @@ go_test(
     ],
     embed = [":server"],
     deps = [
-        "//agent/internal/envoy",
+        "//agent/internal/envoy/admin",
         "//agent/internal/spire",
         "//agent/internal/xds/cache",
         "//agent/pkg/constants",
