+++ agent/internal/envoy/admin/BUILD.bazel
@@ -1,16 +1,19 @@
 load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "envoy",
-    srcs = ["admin.go"],
-    importpath = "github.com/bpalermo/aether/agent/internal/envoy",
+    name = "admin",
+    srcs = [
+        "admin.go",
+        "config_dump.go",
+    ],
+    importpath = "github.com/bpalermo/aether/agent/internal/envoy/admin",
     visibility = ["//agent:__subpackages__"],
 )
 
 go_test(
-    name = "envoy_test",
+    name = "admin_test",
     srcs = ["admin_test.go"],
-    embed = [":envoy"],
+    embed = [":admin"],
     deps = [
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
